♻️ refactor: move AI providers to plugins/ with init() self-registration
Move anthropic, openai, openai-response provider packages from
internal/ai/providers/ to plugins/providers/ and add init()-based
self-registration with the plugin provider registry. Remove the
now-unnecessary register_builtins.go.
