+++ internal/admin/providers.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/vaayne/anna/internal/ai"
-	"github.com/vaayne/anna/internal/ai/providers/anthropic"
-	"github.com/vaayne/anna/internal/ai/providers/openai"
-	openairesponse "github.com/vaayne/anna/internal/ai/providers/openai-response"
+	"github.com/vaayne/anna/plugins/providers/anthropic"
+	"github.com/vaayne/anna/plugins/providers/openai"
+	openairesponse "github.com/vaayne/anna/plugins/providers/openai-response"
 	"github.com/vaayne/anna/internal/config"
 )
