+++ internal/agent/integration_tool_test.go
@@ -11,9 +11,9 @@ import (
 
 	"github.com/vaayne/anna/internal/agent/engine"
 	"github.com/vaayne/anna/internal/ai"
-	"github.com/vaayne/anna/internal/ai/providers/anthropic"
-	"github.com/vaayne/anna/internal/ai/providers/openai"
-	openairesponse "github.com/vaayne/anna/internal/ai/providers/openai-response"
+	"github.com/vaayne/anna/plugins/providers/anthropic"
+	"github.com/vaayne/anna/plugins/providers/openai"
+	openairesponse "github.com/vaayne/anna/plugins/providers/openai-response"
 )
 
 func skipWithoutAPIKey(t *testing.T) string {
