+++ plugins/providers/anthropic/options_test.go
