+++ plugins/providers/anthropic/stream.go
