+++ plugins/providers/anthropic/stream_test.go
