+++ plugins/providers/openai-response/client_test.go
