+++ plugins/providers/openai-response/convert_messages.go
