+++ plugins/providers/openai-response/options.go
