+++ plugins/providers/openai-response/options_test.go
