+++ plugins/providers/openai-response/stream_test.go
