+++ plugins/providers/openai/client_test.go
