+++ plugins/providers/openai/options.go
