+++ plugins/providers/openai/options_test.go
