+++ plugins/providers/openai/stream.go
