+++ plugins/providers/openai/stream_test.go
